+++ client/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LobbyComponent } from '../modules/lobby/lobby.component';
+import { LobbyComponent } from '../components/lobby/lobby.component';
 
 const routes: Routes = [
   {
