--- client/src/modules/lobby/lobby.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { LobbyComponent } from './lobby.component';
-
-@NgModule({
-  declarations: [LobbyComponent],
-  imports: [
-    CommonModule
-  ],
-  providers: [],
-  exports: [LobbyComponent]
-})
-export class LobbyModule { }
